Initial implementation of undo & redo checkpointing in TextArea
