+++ src/textual/widgets/_text_area.py
@@ -3,14 +3,14 @@ from __future__ import annotations
 import dataclasses
 import re
 from collections import defaultdict
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from functools import lru_cache
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, Iterable, Optional, Tuple
+from typing import TYPE_CHECKING, ClassVar, Iterable, Optional, Sequence, Tuple
 
 from rich.style import Style
 from rich.text import Text
-from typing_extensions import Literal, Protocol, runtime_checkable
+from typing_extensions import Literal
 
 from textual._text_area_theme import TextAreaTheme
 from textual._tree_sitter import TREE_SITTER
@@ -24,6 +24,7 @@ from textual.document._document import (
     _utf8_encode,
 )
 from textual.document._document_navigator import DocumentNavigator
+from textual.document._edit import Edit
 from textual.document._history import EditHistory
 from textual.document._languages import BUILTIN_LANGUAGES
 from textual.document._syntax_aware_document import (
@@ -1210,36 +1211,68 @@ TextArea:light .text-area--cursor {
         self.post_message(self.Changed(self))
         return result
 
-    def undo(self) -> EditResult | None:
-        """Undo the most recent edit."""
+    def undo(self) -> None:
+        """Undo the edits since the last checkpoint (the most recent batch of edits)."""
+        edits = self._edit_history.pop_undo()
+        self._undo_batch(edits)
 
-        if self._edit_history:
-            original_edit = self._edit_history.pop()
+    def redo(self) -> None:
+        edits = self._edit_history.pop_redo()
+        self._redo_batch(edits)
 
-            old_gutter_width = self.gutter_width
-            undo_result = original_edit.undo(self)
-            new_gutter_width = self.gutter_width
+    def _undo_batch(self, edits: Sequence[Edit]) -> None:
+        """Undo a batch of Edits."""
+        if not edits:
+            return
 
-            if old_gutter_width != new_gutter_width:
-                self.wrapped_document.wrap(self.wrap_width, self.indent_width)
-            else:
-                self.wrapped_document.wrap_range(
-                    original_edit.from_location,
-                    original_edit._edit_result.end_location,
-                    original_edit.to_location,
-                )
+        old_gutter_width = self.gutter_width
+        for edit in reversed(edits):
+            edit.undo(self)
+        new_gutter_width = self.gutter_width
 
-            self._refresh_size()
-            original_edit.after(self)
-            self._build_highlight_map()
-            self.post_message(self.Changed(self))
-            self._redo_stack.append(original_edit)
-            return undo_result
+        if old_gutter_width != new_gutter_width:
+            self.wrapped_document.wrap(self.wrap_width, self.indent_width)
+        else:
+            # TODO - inefficient
+            self.wrapped_document.wrap_range(
+                min(edit.from_location for edit in edits),
+                max(edit._edit_result.end_location for edit in edits),
+                max(edit.to_location for edit in edits),
+            )
 
-    def redo(self) -> EditResult | None:
-        if self._redo_stack:
-            original_edit = self._redo_stack.pop()
-            return self.edit(original_edit)
+        self._refresh_size()
+        for edit in reversed(edits):
+            edit.after(self)
+        self._build_highlight_map()
+        self.post_message(self.Changed(self))
+
+    def _redo_batch(self, edits: Sequence[Edit]) -> None:
+        """Perform a batch of Edits."""
+        if not edits:
+            return
+
+        old_gutter_width = self.gutter_width
+        edit_results = []
+        for edit in edits:
+            result = edit.do(self)
+            edit_results.append(result)
+        new_gutter_width = self.gutter_width
+
+        if old_gutter_width != new_gutter_width:
+            self.wrapped_document.wrap(self.wrap_width, self.indent_width)
+        else:
+            # TODO - inefficient
+            self.wrapped_document.wrap_range(
+                min(edit.from_location for edit in edits),
+                max(edit.to_location for edit in edits),
+                max(result.end_location for result in edit_results),
+            )
+
+        self._refresh_size()
+        for edit in edits:
+            edit.after(self)
+        self._build_highlight_map()
+        self.post_message(self.Changed(self))
 
     async def _on_key(self, event: events.Key) -> None:
         """Handle key presses which correspond to document inserts."""
@@ -1984,158 +2017,6 @@ TextArea:light .text-area--cursor {
             to_location = (cursor_row, current_row_length)
 
         self.delete(end, to_location, maintain_selection_offset=False)
-
-
-@dataclass
-class Edit:
-    """Implements the Undoable protocol to replace text at some range within a document."""
-
-    text: str
-    """The text to insert. An empty string is equivalent to deletion."""
-    from_location: Location
-    """The start location of the insert."""
-    to_location: Location
-    """The end location of the insert"""
-    maintain_selection_offset: bool
-    """If True, the selection will maintain its offset to the replacement range."""
-    _updated_selection: Selection | None = field(init=False, default=None)
-    """Where the selection should move to after the replace happens."""
-    _edit_result: EditResult | None = field(init=False, default=None)
-    """The result of doing the edit."""
-
-    def do(self, text_area: TextArea) -> EditResult:
-        """Perform the edit operation.
-
-        Args:
-            text_area: The `TextArea` to perform the edit on.
-
-        Returns:
-            An `EditResult` containing information about the replace operation.
-        """
-        text = self.text
-
-        edit_from = self.from_location
-        edit_to = self.to_location
-
-        # This code is mostly handling how we adjust TextArea.selection
-        # when an edit is made to the document programmatically.
-        # We want a user who is typing away to maintain their relative
-        # position in the document even if an insert happens before
-        # their cursor position.
-
-        edit_top, edit_bottom = sorted((edit_from, edit_to))
-        edit_bottom_row, edit_bottom_column = edit_bottom
-
-        selection_start, selection_end = text_area.selection
-        selection_start_row, selection_start_column = selection_start
-        selection_end_row, selection_end_column = selection_end
-
-        edit_result = text_area.document.replace_range(edit_from, edit_to, text)
-
-        new_edit_to_row, new_edit_to_column = edit_result.end_location
-
-        column_offset = new_edit_to_column - edit_bottom_column
-        target_selection_start_column = (
-            selection_start_column + column_offset
-            if edit_bottom_row == selection_start_row
-            and edit_bottom_column <= selection_start_column
-            else selection_start_column
-        )
-        target_selection_end_column = (
-            selection_end_column + column_offset
-            if edit_bottom_row == selection_end_row
-            and edit_bottom_column <= selection_end_column
-            else selection_end_column
-        )
-
-        row_offset = new_edit_to_row - edit_bottom_row
-        target_selection_start_row = selection_start_row + row_offset
-        target_selection_end_row = selection_end_row + row_offset
-
-        if self.maintain_selection_offset:
-            self._updated_selection = Selection(
-                start=(target_selection_start_row, target_selection_start_column),
-                end=(target_selection_end_row, target_selection_end_column),
-            )
-        else:
-            self._updated_selection = Selection.cursor(edit_result.end_location)
-
-        self._edit_result = edit_result
-        return edit_result
-
-    def undo(self, text_area: TextArea) -> EditResult:
-        """Undo the edit operation.
-
-        Looks at the data stored in the edit, and performs the inverse operation of `Edit.do`.
-
-        Args:
-            text_area: The `TextArea` to undo the insert operation on.
-
-        Returns:
-            An `EditResult` containing information about the replace operation.
-        """
-        # This is where the selection will be updated to after the content is restored.
-        target_from = self.from_location
-        target_to = self.to_location
-
-        target_top, target_bottom = sorted((target_to, target_from))
-
-        # The text that was there before and is no longer there - needs to be inserted again.
-        replaced_text = self._edit_result.replaced_text
-
-        # The bounds of the new content
-        # target_from -> edit_result.new_end
-        edit_end = self._edit_result.end_location
-
-        # Replace the span of the edit with the text that was originally there.
-        undo_edit_result = text_area.document.replace_range(
-            target_top, edit_end, replaced_text
-        )
-
-        # TODO - this should be a separate field
-        self._updated_selection = Selection(target_from, target_to)
-
-        return undo_edit_result
-
-    def after(self, text_area: TextArea) -> None:
-        """Possibly update the cursor location after the widget has been refreshed.
-
-        Args:
-            text_area: The `TextArea` this operation was performed on.
-        """
-        if self._updated_selection is not None:
-            text_area.selection = self._updated_selection
-        text_area.record_cursor_width()
-
-
-@runtime_checkable
-class Undoable(Protocol):
-    """Protocol for actions performed in the text editor which can be done and undone.
-
-    These are typically actions which affect the document (e.g. inserting and deleting
-    text), but they can really be anything.
-
-    To perform an edit operation, pass the Edit to `TextArea.edit()`"""
-
-    def do(self, text_area: TextArea) -> Any:
-        """Do the action.
-
-        Args:
-            The `TextArea` to perform the action on.
-
-        Returns:
-            Anything. This protocol doesn't prescribe what is returned.
-        """
-
-    def undo(self, text_area: TextArea) -> Any:
-        """Undo the action.
-
-        Args:
-            The `TextArea` to perform the action on.
-
-        Returns:
-            Anything. This protocol doesn't prescribe what is returned.
-        """
 
 
 @lru_cache(maxsize=128)
