+++ src/textual/widgets/text_area.py
@@ -7,11 +7,11 @@ from textual.document._document import (
     Selection,
 )
 from textual.document._document_navigator import DocumentNavigator
+from textual.document._edit import Edit
 from textual.document._languages import BUILTIN_LANGUAGES
 from textual.document._syntax_aware_document import SyntaxAwareDocument
 from textual.document._wrapped_document import WrappedDocument
 from textual.widgets._text_area import (
-    Edit,
     EndColumn,
     Highlight,
     HighlightName,
